Move template one level up
